chore: split type definitions in multiple files
