+++ src/routes/+layout.ts
@@ -1,1 +1,2 @@
 export const prerender = true;
+export const csr = false;
